.archieve/v2/app/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
dnspython==2.7.0
filetype==1.2.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.1.9
langchain-text-splitters==0.3.9
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.2
langsmith==0.4.15
orjson==3.11.2
ormsgpack==1.10.0
packaging==25.0
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-core==2.33.2
pymongo==4.14.1
python-dotenv==1.1.1
python-telegram-bot==22.3
pytz==2025.2
pyyaml==6.0.2
rapidfuzz==3.13.0
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
schedule==1.2.2
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.43
tenacity==9.1.2
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
xxhash==3.5.0
zstandard==0.24.0
